SC-7761 - Implement changes to support collections.
This ticket introduces the following new information: (a) relation, which defines the relationship between part and children elements of a collection, (b) aggregation level, which should be 2 for parent elements and 1 for children elements, and (c) searchable, which should be 1 for parent elements and 0 for children elements. In particular, the changes in this ticket:

SC-8810 - Implement changes for the rest spiders (Merlin and OEH_spider).
Adding aggregationLevel, searchable, and an empty relation attribute is provided by default (if we pass no value at all from the spiders).

SC-8009: Changed spider version.
SC-8009 - Implement changes to support collections in Mediothek.

1. Relation is now part of LomBase (items.py).
2. LomBase expects a method to include Relation information (lom_base.py).
3. The final transformation that is sent to Edu-Sharing now expects this extra information (relation, aggregation level, and searchable) and passes it (es_connector.py).
4. All the necessary groupings of items into collections and the according selection of parent elements (mediothek_pixiothek_spider.py). Special treatment is done to singular collection elements which describe a full collection.
5. Default values are  set up for the rest spiders (merlin_spider.py and oeh_spider.py).